requirements.txt
openai>=1.0.0
aiofiles>=23.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.0.0
rich>=14.0.0
structlog>=24.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0